Add files via upload
Objective:
To categorize countries using socio-economic and health factors that determine the country's overall development.
Problem Statement:
HELP International has been able to raise $100 million. The CEO wants to use this money strategically and effectively and wants to fund programs in the countries that are most in need of aid. Your job is to categorize the countries using some socio-economic and health factors that determine the overall development of the country. In the end, you need to suggest how to allocate the funds among countries.